requirements.txt
autopep8==1.3.3
bytecode==0.5
httplib2==0.11.3
logzero==1.5.0
neo-boa==0.3.7
pkg-resources==0.0.0
pycodestyle==2.3.1
